requirements.txt
Django>=5.0.4,<5.1
djangorestframework>=3.15.1,<3.16
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.2,<0.28
pillow>=10.3.0,<10.4.1



